+++ backend-rust/CHANGELOG.md
@@ -2,9 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.1.25] - 2025-02-14
-
 ## [Unreleased]
+
+Database schema version: 5
+
+### Fixed
+
+- Add database migration fixing:
+  - Invalid bakers caused by `DelegationEvent::RemoveBaker` event not being handled by the indexer until now.
+  - Invalid delegator state, caused by validator/baker getting removed or changing status to 'ClosedForAll' without moving delegators to the passive pool.
+- Fixed indexer missing handling of moving delegators as pool got removed or closed.
+- Fixed indexer missing handling of event of baker switching directly to delegation.
+
+## [0.1.25] - 2025-02-14
 
 Database schema version: 4